Extract RandomSampler from RandomIndexer
